remove redundant cast

+++ GLPM.TechnicalAnalysis/TA/TACore.cs
@@ -156,7 +156,7 @@ namespace GLPM.TechnicalAnalysis
                     tempReal += inReal_0[today];
                     today++;
                 }
-                prevMA = tempReal / ((double)optInTimePeriod_0);
+                prevMA = tempReal / optInTimePeriod_0;
             }
             while (today <= startIdx)
             {
@@ -215,7 +215,7 @@ namespace GLPM.TechnicalAnalysis
                     tempReal += inReal_0[today];
                     today++;
                 }
-                prevMA = tempReal / ((double)optInTimePeriod_0);
+                prevMA = tempReal / optInTimePeriod_0;
             }
             while (today <= startIdx)
             {
@@ -256,7 +256,7 @@ namespace GLPM.TechnicalAnalysis
             }
             if (optInSlowPeriod_1 != 0)
             {
-                k1 = 2.0 / ((double)(optInSlowPeriod_1 + 1));
+                k1 = 2.0 / (optInSlowPeriod_1 + 1);
             }
             else
             {
@@ -265,7 +265,7 @@ namespace GLPM.TechnicalAnalysis
             }
             if (optInFastPeriod_0 != 0)
             {
-                k2 = 2.0 / ((double)(optInFastPeriod_0 + 1));
+                k2 = 2.0 / (optInFastPeriod_0 + 1);
             }
             else
             {
@@ -326,7 +326,7 @@ namespace GLPM.TechnicalAnalysis
                 fastEMABuffer[i] -= slowEMABuffer[i];
             }
             Array.Copy(fastEMABuffer, lookbackSignal, outMACD_0, 0, (endIdx - startIdx) + 1);
-            retCode = TA_INT_EMA(0, outNbElement1 - 1, fastEMABuffer, optInSignalPeriod_2, 2.0 / ((double)(optInSignalPeriod_2 + 1)), ref outBegIdx2, ref outNbElement2, outMACDSignal_1);
+            retCode = TA_INT_EMA(0, outNbElement1 - 1, fastEMABuffer, optInSignalPeriod_2, 2.0 / (optInSignalPeriod_2 + 1), ref outBegIdx2, ref outNbElement2, outMACDSignal_1);
             if (retCode != RetCode.Success)
             {
                 outBegIdx = 0;
@@ -359,7 +359,7 @@ namespace GLPM.TechnicalAnalysis
             }
             if (optInSlowPeriod_1 != 0)
             {
-                k1 = 2.0 / ((double)(optInSlowPeriod_1 + 1));
+                k1 = 2.0 / (optInSlowPeriod_1 + 1);
             }
             else
             {
@@ -368,7 +368,7 @@ namespace GLPM.TechnicalAnalysis
             }
             if (optInFastPeriod_0 != 0)
             {
-                k2 = 2.0 / ((double)(optInFastPeriod_0 + 1));
+                k2 = 2.0 / (optInFastPeriod_0 + 1);
             }
             else
             {
@@ -429,7 +429,7 @@ namespace GLPM.TechnicalAnalysis
                 fastEMABuffer[i] -= slowEMABuffer[i];
             }
             Array.Copy(fastEMABuffer, lookbackSignal, outMACD_0, 0, (endIdx - startIdx) + 1);
-            retCode = TA_INT_EMA(0, outNbElement1 - 1, fastEMABuffer, optInSignalPeriod_2, 2.0 / ((double)(optInSignalPeriod_2 + 1)), ref outBegIdx2, ref outNbElement2, outMACDSignal_1);
+            retCode = TA_INT_EMA(0, outNbElement1 - 1, fastEMABuffer, optInSignalPeriod_2, 2.0 / (optInSignalPeriod_2 + 1), ref outBegIdx2, ref outNbElement2, outMACDSignal_1);
             if (retCode != RetCode.Success)
             {
                 outBegIdx = 0;
@@ -598,7 +598,7 @@ namespace GLPM.TechnicalAnalysis
                 double tempReal = periodTotal;
                 periodTotal -= inReal_0[trailingIdx];
                 trailingIdx++;
-                outReal_0[outIdx] = tempReal / ((double)optInTimePeriod_0);
+                outReal_0[outIdx] = tempReal / optInTimePeriod_0;
                 outIdx++;
             }
             while (i <= endIdx);
@@ -638,7 +638,7 @@ namespace GLPM.TechnicalAnalysis
                 double tempReal = periodTotal;
                 periodTotal -= inReal_0[trailingIdx];
                 trailingIdx++;
-                outReal_0[outIdx] = tempReal / ((double)optInTimePeriod_0);
+                outReal_0[outIdx] = tempReal / optInTimePeriod_0;
                 outIdx++;
             }
             while (i <= endIdx);
@@ -665,7 +665,7 @@ namespace GLPM.TechnicalAnalysis
                 tempReal = inReal[endSum];
                 tempReal *= tempReal;
                 periodTotal2 += tempReal;
-                double meanValue2 = periodTotal2 / ((double)timePeriod);
+                double meanValue2 = periodTotal2 / timePeriod;
                 tempReal = inReal[startSum];
                 tempReal *= tempReal;
                 periodTotal2 -= tempReal;
@@ -704,7 +704,7 @@ namespace GLPM.TechnicalAnalysis
                 tempReal = inReal[endSum];
                 tempReal *= tempReal;
                 periodTotal2 += tempReal;
-                double meanValue2 = periodTotal2 / ((double)timePeriod);
+                double meanValue2 = periodTotal2 / timePeriod;
                 tempReal = inReal[startSum];
                 tempReal *= tempReal;
                 periodTotal2 -= tempReal;
@@ -761,8 +761,8 @@ namespace GLPM.TechnicalAnalysis
                 periodTotal1 += tempReal;
                 tempReal *= tempReal;
                 periodTotal2 += tempReal;
-                double meanValue1 = periodTotal1 / ((double)optInTimePeriod_0);
-                double meanValue2 = periodTotal2 / ((double)optInTimePeriod_0);
+                double meanValue1 = periodTotal1 / optInTimePeriod_0;
+                double meanValue2 = periodTotal2 / optInTimePeriod_0;
                 tempReal = inReal_0[trailingIdx];
                 trailingIdx++;
                 periodTotal1 -= tempReal;
@@ -813,8 +813,8 @@ namespace GLPM.TechnicalAnalysis
                 periodTotal1 += tempReal;
                 tempReal *= tempReal;
                 periodTotal2 += tempReal;
-                double meanValue1 = periodTotal1 / ((double)optInTimePeriod_0);
-                double meanValue2 = periodTotal2 / ((double)optInTimePeriod_0);
+                double meanValue1 = periodTotal1 / optInTimePeriod_0;
+                double meanValue2 = periodTotal2 / optInTimePeriod_0;
                 tempReal = inReal_0[trailingIdx];
                 trailingIdx++;
                 periodTotal1 -= tempReal;
