+++ GLPM.TechnicalAnalysis/TA/TAFunc/TA_Asin.cs
@@ -78,7 +78,7 @@ namespace GLPM.TechnicalAnalysis
             int outIdx = 0;
             while (i <= endIdx)
             {
-                outReal[outIdx] = Math.Asin((double)inReal[i]);
+                outReal[outIdx] = Math.Asin(inReal[i]);
                 i++;
                 outIdx++;
             }
