+++ GLPM.TechnicalAnalysis/TA/TAFunc/TA_Cmo.cs
@@ -90,8 +90,8 @@ namespace GLPM.TechnicalAnalysis
                         }
                     }
 
-                    tempValue1 = prevLoss / ((double)optInTimePeriod);
-                    tempValue2 = prevGain / ((double)optInTimePeriod);
+                    tempValue1 = prevLoss / optInTimePeriod;
+                    tempValue2 = prevGain / optInTimePeriod;
                     double tempValue3 = tempValue2 - tempValue1;
                     double tempValue4 = tempValue1 + tempValue2;
                     if ((-1E-08 >= tempValue4) || (tempValue4 >= 1E-08))
@@ -135,8 +135,8 @@ namespace GLPM.TechnicalAnalysis
                     }
                 }
 
-                prevLoss /= (double)optInTimePeriod;
-                prevGain /= (double)optInTimePeriod;
+                prevLoss /= optInTimePeriod;
+                prevGain /= optInTimePeriod;
                 if (today > startIdx)
                 {
                     tempValue1 = prevGain + prevLoss;
@@ -169,8 +169,8 @@ namespace GLPM.TechnicalAnalysis
                             prevGain += tempValue2;
                         }
 
-                        prevLoss /= (double)optInTimePeriod;
-                        prevGain /= (double)optInTimePeriod;
+                        prevLoss /= optInTimePeriod;
+                        prevGain /= optInTimePeriod;
                         today++;
                     }
                 }
@@ -192,8 +192,8 @@ namespace GLPM.TechnicalAnalysis
                         prevGain += tempValue2;
                     }
 
-                    prevLoss /= (double)optInTimePeriod;
-                    prevGain /= (double)optInTimePeriod;
+                    prevLoss /= optInTimePeriod;
+                    prevGain /= optInTimePeriod;
                     tempValue1 = prevGain + prevLoss;
                     if ((-1E-08 >= tempValue1) || (tempValue1 >= 1E-08))
                     {
@@ -300,8 +300,8 @@ namespace GLPM.TechnicalAnalysis
                         }
                     }
 
-                    tempValue1 = prevLoss / ((double)optInTimePeriod);
-                    tempValue2 = prevGain / ((double)optInTimePeriod);
+                    tempValue1 = prevLoss / optInTimePeriod;
+                    tempValue2 = prevGain / optInTimePeriod;
                     double tempValue3 = tempValue2 - tempValue1;
                     double tempValue4 = tempValue1 + tempValue2;
                     if ((-1E-08 >= tempValue4) || (tempValue4 >= 1E-08))
@@ -345,8 +345,8 @@ namespace GLPM.TechnicalAnalysis
                     }
                 }
 
-                prevLoss /= (double)optInTimePeriod;
-                prevGain /= (double)optInTimePeriod;
+                prevLoss /= optInTimePeriod;
+                prevGain /= optInTimePeriod;
                 if (today > startIdx)
                 {
                     tempValue1 = prevGain + prevLoss;
@@ -379,8 +379,8 @@ namespace GLPM.TechnicalAnalysis
                             prevGain += tempValue2;
                         }
 
-                        prevLoss /= (double)optInTimePeriod;
-                        prevGain /= (double)optInTimePeriod;
+                        prevLoss /= optInTimePeriod;
+                        prevGain /= optInTimePeriod;
                         today++;
                     }
                 }
@@ -402,8 +402,8 @@ namespace GLPM.TechnicalAnalysis
                         prevGain += tempValue2;
                     }
 
-                    prevLoss /= (double)optInTimePeriod;
-                    prevGain /= (double)optInTimePeriod;
+                    prevLoss /= optInTimePeriod;
+                    prevGain /= optInTimePeriod;
                     tempValue1 = prevGain + prevLoss;
                     if ((-1E-08 >= tempValue1) || (tempValue1 >= 1E-08))
                     {
