+++ GLPM.TechnicalAnalysis/TA/TAFunc/TA_Cos.cs
@@ -78,7 +78,7 @@ namespace GLPM.TechnicalAnalysis
             int outIdx = 0;
             while (i <= endIdx)
             {
-                outReal[outIdx] = Math.Cos((double)inReal[i]);
+                outReal[outIdx] = Math.Cos(inReal[i]);
                 i++;
                 outIdx++;
             }
